Adjust spatial extent and timestep of GIA model
Sensitivity testing showed that the modeled uplift is sensitive to the
spatial extent of the GIA domain.  The spatial extent can be increased
by either making the GIA mesh larger or increasing the 'fac' parameter
that adds padding.  Also, the existing GIA domain was too small and led
to interpolation artifacts so needs to be enlarged.

The interpolation artifacts were remvoed by adding a buffer of blank
space around the MPAS mesh extent in the GIA mesh.
Convergent behavior in uplift rate was achieved when the effective padding factor is
about 5x the MPAS domain.  This commit adds a buffer of 0.2e6 around the
MPAS mesh on all sides (about a 60% increase in domain size over the
MPAS mesh extent) and also changes the fac value from 2 to 3.  Other
combinations would also work, but this combination is committed because
it was tested.  It was demonstrated that with this buffer, the
difference in uplift rate between fac=3 and fac=4 are negligible, but
differences with smaller fac values are significant.

Increasing the GIA domain size makes the model much slower.  The old
setup used no buffer (1.0*MPAS) with fac=2 for an effective domain size
of 2*MPAS.  The new one uses buffer=1.6*MPAS with fac=3 for an effective
size of 5*MPAS.  Accounting for both x and y dimensions, this amounts to
a 25/4=6.25x increase in model cost.  However, it was also determined
that using time steps of 0.01 year is not necessary to maintain
accuracy, and the GIA model time step could be made much longer.  Even
with dt=1 yr, the differences were only slight compared to dt=0.01 yr.
But the dt is adjusted to 0.1 yr to be safe.  The 10x speedup in model
timestep already more than offsets the increased cost from expanding the
spatial extent.